feat: screen agnostic background
Adding the background image and making it screen agnostic. In certain
aspect ratios where resize to fill cuts away part of the vault or
esssential parts like the keyboard/timer, another method is used to
preserve a minimum display area around the center of the background, at
the cost of padding border visible on the screen.