feat[controller]: setup basic user interaction
When the user clicks on the door, depending on the left/right side,
initiate a wheel rotation.

Each time the game resets, the model will initiate a new sequence. The
controller will print that sequence in the console.

For now, on error or success the state reset. Opening the door, or wild
handle rotation is not implemented, yet.