feat[app]: setup service-worker and manifest
Service worker will cache all files/assets necessary to display the
image. Those will not be requested again, unless the index changes.

The manifest allow installation of the game as PWA. On phones this can
add the page as an app on the home screen and start in fullsreen model.
However, without access to the developer console, the correct
combination is too difficult to guess.